common: Move more GUI stuff into the gui module
